requirements.txt
streamlit
streamlit-lottie
phidata 
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
openai
sqlalchemy
pgvector
psycopg[binary]
pypdf
phidata
google-generativeai
duckduckgo-search
pytube